use default libdir (lib / lib64) as UDUNITS2 libdir
Some 64bit distributions use 'lib64' instead of 'lib' as library
directory. The base directory of $libdir is taken (on most distros
it is 'lib' or 'lib64') and appended to $UDUNITS2_PATH to generate
the UDUNITS2 library directory. If this directory does not exist,
the directory is set to $UDUNITS2_PATH/lib.