test file in_4c.nc now is created during make test
The ./data directory contains *.cdl files, from which netCDF files for
the NCO tests are build. The makefile for creating these files is
called in the beginning of the build process. However, creating the
file in_4c.nc needs 'ncks'. All other files are generated via 'ncgen'.
Because ncks has not been compiled in the beginning of the build
process, that file is not generated/created. Two makefiles were
modified:
 - ./data/Makefile: path to the compiled ncks binary (path BEFORE
                    'make install') -> $abs_top_builddir/src/nco/ncks
 - ./Makefile: the 'tst' target now does
                > cd data && $(MAKE) in_4c.nc && cd ..
               to generate the 'in_4c.nc' before the actual tests are
               performed

The Makefile.in and Makefile.am were equally modified (one doesn't
need to newly generate the Makefile.in files).)